move packages around
