+++ src/main/java/org/maxgamer/rs/assets/codec/asset/EncryptedException.java
@@ -1,4 +1,4 @@
-package org.maxgamer.rs.cache;
+package org.maxgamer.rs.assets.codec.asset;
 
 import java.io.IOException;
