+++ src/main/java/org/maxgamer/rs/assets/protocol/format/CS2.java
@@ -1,4 +1,4 @@
-package org.maxgamer.rs.cache.format;
+package org.maxgamer.rs.assets.protocol.format;
 
 import org.maxgamer.rs.structure.ReflectUtil;
