+++ src/main/java/org/maxgamer/rs/assets/protocol/format/CS2Factory.java
@@ -1,4 +1,4 @@
-package org.maxgamer.rs.cache.format;
+package org.maxgamer.rs.assets.protocol.format;
 
 import java.nio.ByteBuffer;
