+++ src/main/java/org/maxgamer/rs/assets/protocol/format/GameObjectFactory.java
@@ -1,4 +1,4 @@
-package org.maxgamer.rs.cache.format;
+package org.maxgamer.rs.assets.protocol.format;
 
 import org.maxgamer.rs.model.map.object.GameObjectProto;
