+++ src/main/java/org/maxgamer/rs/assets/protocol/format/Landscape.java
@@ -1,4 +1,4 @@
-package org.maxgamer.rs.cache.format;
+package org.maxgamer.rs.assets.protocol.format;
 
 import org.maxgamer.rs.model.map.*;
 import org.maxgamer.rs.model.map.object.GameObject;
