+++ src/main/java/org/maxgamer/rs/model/map/DynamicMap.java
@@ -1,8 +1,8 @@
 package org.maxgamer.rs.model.map;
 
-import org.maxgamer.rs.cache.EncryptedException;
-import org.maxgamer.rs.cache.MapCache;
-import org.maxgamer.rs.cache.format.Landscape;
+import org.maxgamer.rs.assets.codec.asset.EncryptedException;
+import org.maxgamer.rs.assets.protocol.MapCache;
+import org.maxgamer.rs.assets.protocol.format.Landscape;
 import org.maxgamer.rs.model.events.world.ChunkLoadEvent;
 
 import java.io.FileNotFoundException;
